Meets: support records
For records, we need to find the union of their field names, and for each
name in the resulting set, assign it the meet of the two original fields'
types.

Happily, Hash#merge takes a block which is used to resolve cases where the
value for the same key is different in each hash, which is very convenient
here.