Subtyping: SA-RECD (permutation & width cases)
Note that we had to "fix" equality on the Record type. This is because `#==`
on structs is defined by comparing the struct's fields. For Record, the only
field is a Hash, and equality for Hash ignores field order. However, we want
the two types {x:Bool, y:Bool} and {y:Bool, x:Bool} to be considered distinct.