Joins: support joining records
The join of two records is the intersection of their fields, with the type of
each member of the intersection being the join of the original fields' types.

E.g. if the two records being joined have fields `x:Bool` and `x:Bool -> Bool`
respectively, the join should have a field `x:Top`.